refactor: move exported types to /@api Signed-off-by: Philippe Martin <phmartin@redhat.com>
